requirement.txt
streamlit>=1.22.0
ultralytics>=8.0.0
opencv-python>=4.7.0
Pillow>=9.0.0
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
torch>=1.13.0